Clean up, reorganise folders/imports
