--- web/src/app/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import petReducer from '../components/PetList/reducers';
-
-export const store = configureStore({
-  reducer: {
-    pets: petReducer,
-  },
-});
