+++ web/src/components/AddItem/index.js
@@ -1,0 +1,3 @@
+import AddItem from './AddItem';
+
+export default AddItem;
