+++ web/src/components/App/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import logo from './logo.svg';
-import PetList from './components/PetList/';
+import logo from '../../assets/logo.svg';
+import PetList from '../PetList';
 import './App.css';
 
 const App = () => (
@@ -8,9 +8,9 @@ const App = () => (
     <header className="App-header">
       <img src={logo} className="App-logo" alt="logo" />
     </header>
-      <div className="App-content">
-        <PetList data-testid="pet-list" />
-      </div>
+    <div className="App-content">
+      <PetList data-testid="pet-list" />
+    </div>
   </div>
 );
